requirements.txt
python-dotenv
langchain-groq
redis
fastapi[standard]
uvicorn